requirements.txt
Flask
Flask-SocketIO
flask-testing
pytest
pytest-flask
pytest-mock
openai==0.28
python-dotenv
coverage
eventlet 